# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Taiwan Tourism Open Data API
  description: '

    以 **交通部觀光署「景點 - 觀光資訊資料庫」** 政府開放資料（6,000+ 筆全台景點）為基礎的 RESTful API。


    資料來源：<https://data.gov.tw/dataset/7777>（觀光資料標準 V2.1）


    ### 資源 (Resources)

    | Resource | 說明 |

    |---|---|

    | `attractions` | 景點（支援關鍵字搜尋、多條件篩選、排序、分頁、附近景點查詢） |

    | `attractions/{id}/images` | 景點圖片（子資源） |

    | `categories` | 景點類型（文化類、生態類…） |

    | `cities` / `towns` | 縣市 / 鄉鎮市區 |

    | `reviews` | 遊客評論（1~5 星） |

    | `trips` / `trips/{id}/items` | 旅遊行程與行程項目 |

    | `stats` | 統計資訊（唯讀） |


    ### 驗證

    所有 **讀取 (GET)** 操作公開；**新增 / 修改 / 刪除 (POST / PUT / PATCH / DELETE)** 需在 Header
    帶入

    `X-API-Key: ntub-iot-2026`（可按右上角 **Authorize** 按鈕設定）。


    ### 慣例

    * 列表回應格式：`{"total", "page", "page_size", "pages", "items"}`，並附 `X-Total-Count`
    header

    * 建立成功回傳 `201 Created` 與 `Location` header；刪除成功回傳 `204 No Content`

    * 錯誤格式：`{"detail": "錯誤說明"}`（404 查無資料、409 資料衝突、422 驗證失敗）

    '
  contact:
    name: NTUB IoT App 作業
  license:
    name: 政府資料開放授權條款－第1版
    url: https://data.gov.tw/license
  version: 1.0.0
paths:
...
  /api/v1/trips:
    get:
      tags:
      - Trips 旅遊行程
      summary: 列出旅遊行程
      operationId: list_trips_api_v1_trips_get
      parameters:
      - name: owner
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 建立者
          title: Owner
        description: 建立者
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 標題關鍵字
          title: Q
        description: 標題關鍵字
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 頁碼（從 1 開始）
          default: 1
          title: Page
        description: 頁碼（從 1 開始）
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 每頁筆數（1~100）
          default: 20
          title: Page Size
        description: 每頁筆數（1~100）
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_TripSummary_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Trips 旅遊行程
      summary: 建立行程（可同時帶入行程項目）
      operationId: create_trip_api_v1_trips_post
      security:
      - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trips/{trip_id}:
    get:
      tags:
      - Trips 旅遊行程
      summary: 取得行程（含所有行程項目）
      operationId: get_trip_api_v1_trips__trip_id__get
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: integer
          title: Trip Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Trips 旅遊行程
      summary: 整筆更新行程（行程項目會以 body 內容整批取代）
      operationId: replace_trip_api_v1_trips__trip_id__put
      security:
      - APIKeyHeader: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: integer
          title: Trip Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Trips 旅遊行程
      summary: 部分更新行程
      operationId: update_trip_api_v1_trips__trip_id__patch
      security:
      - APIKeyHeader: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: integer
          title: Trip Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Trips 旅遊行程
      summary: 刪除行程
      operationId: delete_trip_api_v1_trips__trip_id__delete
      security:
      - APIKeyHeader: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: integer
          title: Trip Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trips/{trip_id}/items:
    get:
      tags:
      - Trip Items 行程項目
      summary: 列出行程項目
      operationId: list_items_api_v1_trips__trip_id__items_get
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: integer
          title: Trip Id
      - name: day
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          description: 只列出第幾天
          title: Day
        description: 只列出第幾天
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TripItemOut'
                title: Response List Items Api V1 Trips  Trip Id  Items Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Trip Items 行程項目
      summary: 新增行程項目（加入一個景點）
      operationId: create_item_api_v1_trips__trip_id__items_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: integer
          title: Trip Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripItemCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripItemOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trips/{trip_id}/items/{item_id}:
    get:
      tags:
      - Trip Items 行程項目
      summary: 取得單一行程項目
      operationId: get_item_api_v1_trips__trip_id__items__item_id__get
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: integer
          title: Trip Id
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripItemOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Trip Items 行程項目
      summary: 整筆更新行程項目
      operationId: replace_item_api_v1_trips__trip_id__items__item_id__put
      security:
      - APIKeyHeader: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: integer
          title: Trip Id
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripItemCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripItemOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Trip Items 行程項目
      summary: 部分更新行程項目
      operationId: update_item_api_v1_trips__trip_id__items__item_id__patch
      security:
      - APIKeyHeader: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: integer
          title: Trip Id
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripItemUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripItemOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Trip Items 行程項目
      summary: 刪除行程項目
      operationId: delete_item_api_v1_trips__trip_id__items__item_id__delete
      security:
      - APIKeyHeader: []
      parameters:
      - name: trip_id
        in: path
        required: true
        schema:
          type: integer
          title: Trip Id
      - name: item_id
        in: path
        required: true
        schema:
          type: integer
          title: Item Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AttractionSummary:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        city:
          anyOf:
          - $ref: '#/components/schemas/CityRef'
          - type: 'null'
        town:
          anyOf:
          - $ref: '#/components/schemas/TownRef'
          - type: 'null'
        categories:
          items:
            $ref: '#/components/schemas/CategoryRef'
          type: array
          title: Categories
          default: []
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        is_free:
          type: boolean
          title: Is Free
        service_status:
          type: integer
          title: Service Status
        cover_image:
          anyOf:
          - type: string
          - type: 'null'
          title: Cover Image
      type: object
      required:
      - id
      - name
      - is_free
      - service_status
      title: AttractionSummary
...
    CategoryRef:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
      - id
      - name
      title: CategoryRef
...
    CityRef:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
      type: object
      required:
      - code
      - name
      title: CityRef
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Page_TripSummary_:
      properties:
        total:
          type: integer
          title: Total
          description: 符合條件的總筆數
        page:
          type: integer
          title: Page
          description: 目前頁碼（從 1 開始）
        page_size:
          type: integer
          title: Page Size
          description: 每頁筆數
        pages:
          type: integer
          title: Pages
          description: 總頁數
        items:
          items:
            $ref: '#/components/schemas/TripSummary'
          type: array
          title: Items
      type: object
      required:
      - total
      - page
      - page_size
      - pages
      - items
      title: Page[TripSummary]
...
    TownRef:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
      type: object
      required:
      - code
      - name
      title: TownRef
...
    TripCreate:
      properties:
        title:
          type: string
          maxLength: 100
          minLength: 1
          title: Title
          examples:
          - 宜蘭兩天一夜
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - 太平山賞楓之旅
        owner:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Owner
          examples:
          - 小明
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
          examples:
          - '2026-10-10'
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
          examples:
          - '2026-10-11'
        items:
          items:
            $ref: '#/components/schemas/TripItemCreate'
          type: array
          title: Items
          description: 可同時建立行程項目
      type: object
      required:
      - title
      title: TripCreate
    TripItemCreate:
      properties:
        attraction_id:
          type: string
          title: Attraction Id
          examples:
          - Attraction_345040000G_000001
        day:
          type: integer
          maximum: 365.0
          minimum: 1.0
          title: Day
          description: 第幾天
          default: 1
          examples:
          - 1
        sequence:
          type: integer
          maximum: 1000.0
          minimum: 1.0
          title: Sequence
          description: 當天順序
          default: 1
          examples:
          - 1
        note:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Note
          examples:
          - 早上出發看日出
      type: object
      required:
      - attraction_id
      title: TripItemCreate
    TripItemOut:
      properties:
        id:
          type: integer
          title: Id
        trip_id:
          type: integer
          title: Trip Id
        day:
          type: integer
          title: Day
        sequence:
          type: integer
          title: Sequence
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        attraction:
          $ref: '#/components/schemas/AttractionSummary'
      type: object
      required:
      - id
      - trip_id
      - day
      - sequence
      - attraction
      title: TripItemOut
    TripItemUpdate:
      properties:
        attraction_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Attraction Id
        day:
          anyOf:
          - type: integer
            maximum: 365.0
            minimum: 1.0
          - type: 'null'
          title: Day
        sequence:
          anyOf:
          - type: integer
            maximum: 1000.0
            minimum: 1.0
          - type: 'null'
          title: Sequence
        note:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Note
      type: object
      title: TripItemUpdate
    TripOut:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        owner:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
        item_count:
          type: integer
          title: Item Count
          default: 0
        items:
          items:
            $ref: '#/components/schemas/TripItemOut'
          type: array
          title: Items
          default: []
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - title
      title: TripOut
    TripSummary:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        owner:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
        item_count:
          type: integer
          title: Item Count
          default: 0
      type: object
      required:
      - id
      - title
      title: TripSummary
    TripUpdate:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        owner:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Owner
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
      type: object
      title: TripUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      description: 寫入操作需帶入 API Key（預設為 `ntub-iot-2026`，可用環境變數 `API_KEY` 修改）
      in: header
      name: X-API-Key
...
servers:
- url: http://127.0.0.1:8000
  description: 本機開發伺服器